fixed bugs to clean up console output --> adding items multiple times to graphicsscene in controllerareadescriptor --> deleting nonexistent file in DataExporters

+++ BioTracker/CoreApp/BioTracker/Model/DataExporterCSV.cpp
@@ -215,7 +215,7 @@ void DataExporterCSV::writeAll() {
 void DataExporterCSV::close() {
     _ofs.close();
 
-    if (!_root || _root->size() == 0) {
+    if ((!_root || _root->size() == 0) && _tmpFile!="" ) {
         //Remove temporary file
         QFile file(_tmpFile.c_str());
         file.remove();
