+++ BioTracker/CoreApp/BioTracker/Model/DataExporterJson.cpp
@@ -224,7 +224,7 @@ void DataExporterJson::writeAll() {
 void DataExporterJson::close() {
     _ofs.close();
 
-    if (!_root || _root->size() == 0) {
+    if ((!_root || _root->size() == 0) &&_tmpFile!="" ) {
         //Remove temporary file
         QFile file(_tmpFile.c_str());
         file.remove();
