+++ BioTracker/CoreApp/BioTracker/Model/DataExporterSerialize.cpp
@@ -165,7 +165,7 @@ void DataExporterSerialize::writeAll() {
 void DataExporterSerialize::close() {
     _ofs.close();
 
-    if (!_root || _root->size() == 0) {
+    if ((!_root || _root->size() == 0) && _tmpFile!="" ) {
         //Remove temporary file
         QFile file(_tmpFile.c_str());
         file.remove();
